Add overloads to the request invoker that allows the capabilities filter to inspect the parameter to determine whether the command is supported.
WebTools has a way to get this functionality without this change, so it's not needed. I've also changed the default onTypeFormatting to inspect that data, but other default filters could also benefit from this.

I'm not certain this commit improves things, so I'll defer creating a PR until I get feedback indicating whether to proceed.